Comment preview
The comment now previews under the post.

Conditions:
Takes the highest liked comment and presents it under the main comment.
If there is no highest liked comment, then it takes the most recently commented item and presents that to the main page. If there are no comments, nothing is shown under the post.